pyproject.toml
[project]
name = "close-call"
version = "0.1.0"
description = "AI voice agent for practicing German-language solar sales conversations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[google,webrtc]==0.0.107",
    "python-dotenv",
    "google-generativeai",
    "fastapi",
    "uvicorn[standard]",
    "python-frontmatter>=1.1.0",
    "pyinstaller>=6.19.0",
    "requests>=2.33.1",
]

